♻️ refactor: 임시 페이지 코드 삭제
