+++ src/main.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { ChakraProvider } from '@chakra-ui/react';
 import { Global, ThemeProvider } from '@emotion/react';
 import ReactDOM from 'react-dom/client';
@@ -7,12 +6,10 @@ import theme, { chakraTheme } from '@styles/theme';
 import App from 'App';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
-  <React.StrictMode>
-    <ThemeProvider theme={theme}>
-      <ChakraProvider theme={chakraTheme}>
-        <Global styles={globalStyle} />
-        <App />
-      </ChakraProvider>
-    </ThemeProvider>
-  </React.StrictMode>,
+  <ThemeProvider theme={theme}>
+    <ChakraProvider theme={chakraTheme}>
+      <Global styles={globalStyle} />
+      <App />
+    </ChakraProvider>
+  </ThemeProvider>,
 );
